moved db connection to server file, split server into two files - server and app - as per circle ci
