+++ db/dbHelpers.js
@@ -1,7 +1,5 @@
 const mongoose = require('mongoose');
 
-// mongoose.connect('mongodb://localhost/fecproject');
-
 const restaurantSchema = mongoose.Schema({
   id: { type: Number, unique: true },
   name: String,
@@ -27,7 +25,6 @@ const restaurantSchema = mongoose.Schema({
 const Restaurant = mongoose.model('restaurants', restaurantSchema);
 
 const save = (options, cb) => {
-  mongoose.connect('mongodb://localhost/fecproject');
   const { data } = options;
   const Model = options.model;
   let count = 0;
@@ -50,7 +47,6 @@ const save = (options, cb) => {
         count += 1;
         if (count === data.length) {
           cb(result);
-          mongoose.disconnect();
         }
       });
     });
@@ -58,7 +54,6 @@ const save = (options, cb) => {
 };
 
 const find = (options, cb) => {
-  mongoose.connect('mongodb://localhost/fecproject');
   const query = options.query || 'id menu.lunch';
   const idNum = options.id || 90976;
 
@@ -68,7 +63,6 @@ const find = (options, cb) => {
         cb(err, null);
       } else {
         cb(null, data);
-        mongoose.disconnect();
       }
     });
   } else {
@@ -77,7 +71,6 @@ const find = (options, cb) => {
         cb(err, null);
       } else {
         cb(null, data);
-        mongoose.disconnect();
       }
     });
   }
@@ -86,4 +79,5 @@ const find = (options, cb) => {
 module.exports.save = save;
 module.exports.find = find;
 module.exports.Restaurant = Restaurant;
+module.exports.mongoose = mongoose;
