+++ src/spec/db-test/db.test.js
@@ -1,9 +1,13 @@
-const dbHelpers = require('../../../db/dbHelpers');
+const db = require('../../../db/dbHelpers');
+
 
 describe('database helpers - save and find', () => {
+  beforeAll(() => db.mongoose.connect('mongodb://localhost/fecproject'));
+  afterAll(() => db.mongoose.disconnect());
+
   it('should save all data items to test database', (done) => {
     expect.assertions(1);
-    dbHelpers.find({ query: '{}' }, (err, docs) => {
+    db.find({ query: '{}' }, (err, docs) => {
       if (err) { throw err; }
       expect(docs.length).toBe(119);
       done();
@@ -13,7 +17,7 @@ describe('database helpers - save and find', () => {
   // nb: the sample data contains a duplication of the restaurant with id: 89104
   it('should not save duplicate restaurants', async (done) => {
     expect.assertions(2);
-    dbHelpers.find({ id: 89104 }, (err, docs) => {
+    db.find({ id: 89104 }, (err, docs) => {
       if (err) { throw err; }
       expect(docs.length).toBe(1);
       expect(docs[0].id).toBe(89104);
@@ -23,7 +27,7 @@ describe('database helpers - save and find', () => {
 
   it('should return specified data fields only', (done) => {
     expect.assertions(4);
-    dbHelpers.find({ query: 'id menu.dinner' }, (err, docs) => {
+    db.find({ query: 'id menu.dinner' }, (err, docs) => {
       if (err) { throw err; }
       const docsObj = docs[0].toObject();
       expect(docs.length).toBe(1);
